Updating doc
