+++ examples/tutorial/x/blog/handlers_test.go
@@ -208,7 +208,6 @@ func testHandlerDeliver(t *testing.T, testcases []testcase) {
 		}
 	}
 }
-
 func TestCreateBlogMsgHandlerCheck(t *testing.T) {
 	_, signer := x.TestHelpers{}.MakeKey()
 	testHandlerCheck(
@@ -966,113 +965,4 @@ func TestSetProfileMsgHandlerDeliver(t *testing.T) {
 			},
 		},
 	)
-}
-func TestBlogQuery(t *testing.T) {
-	db := store.MemStore()
-	_, signer := helpers.MakeKey()
-	ctx, auth := newContextWithAuth([]weave.Condition{signer})
-	_, err := createBlogMsgHandlerFn(auth).Deliver(ctx, db, newTx(
-		&CreateBlogMsg{
-			Slug:    "this_is_a_blog",
-			Title:   "this is a blog title",
-			Authors: [][]byte{signer.Address()},
-		}))
-	require.NoError(t, err, "failed to deliver blog")
-
-	qr := weave.NewQueryRouter()
-	RegisterQuery(qr)
-
-	h := qr.Handler("/blogs")
-	require.NotNil(t, h)
-
-	blogs, err := h.Query(db, "", []byte("this_is_a_blog"))
-	require.NoError(t, err)
-
-	bucket := NewBlogBucket()
-	expected, err := bucket.Get(db, []byte("this_is_a_blog"))
-	require.NoError(t, err)
-
-	actual, err := bucket.Parse(nil, blogs[0].Value)
-	require.EqualValues(t, expected.Value(), actual.Value())
-}
-func TestPostQuery(t *testing.T) {
-	db := store.MemStore()
-	_, signer := helpers.MakeKey()
-	ctx, auth := newContextWithAuth([]weave.Condition{signer})
-
-	_, err := createBlogMsgHandlerFn(auth).Deliver(ctx, db, newTx(
-		&CreateBlogMsg{
-			Slug:    "this_is_a_blog",
-			Title:   "this is a blog title",
-			Authors: [][]byte{signer.Address()},
-		}))
-	require.NoError(t, err)
-
-	_, err = createPostMsgHandlerFn(auth).Deliver(ctx, db, newTx(
-		&CreatePostMsg{
-			Blog:   "this_is_a_blog",
-			Title:  "this is a post title",
-			Text:   longText,
-			Author: signer.Address(),
-		}))
-	require.NoError(t, err)
-
-	qr := weave.NewQueryRouter()
-	RegisterQuery(qr)
-
-	// query by post
-	h := qr.Handler("/posts")
-	require.NotNil(t, h)
-
-	key := newPostCompositeKey("this_is_a_blog", 1)
-	posts, err := h.Query(db, "", key)
-	require.NoError(t, err)
-
-	bucket := NewPostBucket()
-	expected, err := bucket.Get(db, key)
-	require.NoError(t, err)
-
-	actual, err := bucket.Parse(nil, posts[0].Value)
-	require.EqualValues(t, expected.Value(), actual.Value())
-
-	// query by author
-	h = qr.Handler("/posts/author")
-	require.NotNil(t, h)
-
-	posts, err = h.Query(db, "", signer.Address())
-	require.NoError(t, err)
-
-	expectedSlice, err := bucket.GetIndexed(db, "author", signer.Address())
-	require.NoError(t, err)
-
-	actual, err = bucket.Parse(nil, posts[0].Value)
-	require.EqualValues(t, expectedSlice[0].Value(), actual.Value())
-}
-func TestProfile(t *testing.T) {
-	db := store.MemStore()
-	_, signer := helpers.MakeKey()
-	ctx, auth := newContextWithAuth([]weave.Condition{signer})
-	_, err := SetProfileMsgHandlerFn(auth).Deliver(ctx, db, newTx(
-		&SetProfileMsg{
-			Name:        "lehajam",
-			Description: "my profile description",
-		},
-	))
-	require.NoError(t, err)
-
-	qr := weave.NewQueryRouter()
-	RegisterQuery(qr)
-
-	h := qr.Handler("/profiles")
-	require.NotNil(t, h)
-
-	blogs, err := h.Query(db, "", []byte("lehajam"))
-	require.NoError(t, err)
-
-	bucket := NewProfileBucket()
-	expected, err := bucket.Get(db, []byte("lehajam"))
-	require.NoError(t, err)
-
-	actual, err := bucket.Parse(nil, blogs[0].Value)
-	require.EqualValues(t, expected.Value(), actual.Value())
 }
